select icon function
added a function that can select an icon and then save it to an object in a list. I have a selectedPlayer that will change depending on the player we are on. It starts at 0 and will later get changed when its the next players turn. The iconSelector() will loop thru all the icon buttons and add an eventlistener to each one, if you click on the button it will firstly reset all the borders that are currently there, and then add a border to the one you have clicked. i wanted a border cause its easier to see which one you clicked on at the moment. Then i will set the player 1 icon to the icon you clicked on