Added css + changed what saves in iconSelector()
Added css, changed boxy to no margin or padding since i dont want the white lines, added a low opacity background to #player-selection so i can see everything underneath, also changed so it has position absolute so its ontop of the game-board. (I did this cause i want the players to be able to choose their name and icon first and then go into the game) added some styling to the buttons where the images are, i dont want them to apear as buttons so i added all: unset; which removes the default button styling. Added #game that sets the whole game board and some padding with flex. Then each different div with the ids #player1-score-board, #game-section, #player2-score-board will have everything in its center with flex-direction column and justify content & align items, also wanted flex-grow for them to be equally big, however after some thinking i wanted the middle part (the actual game board) to be a bit bigger so i set it to flex-grow 2.

And now for the js; i only changed so i save the image element that is child to the button instead, with this i can easier showcase the image with only displaying element from the array with the objects later in the game.