sets player name and continue button to next player and the game.
Added a function setPlayerName() to only take the value from the input. With this i can then use an eventListener on the continue button that reacts when you click on it, it will set the player name, and also increment the selectedPlayer. Doing this, i can then "go to the next" player (it still only shows the same images and everything) but instead it now saves it down in the second object in the list of players. Doing this i can now have two players, and next time i click continue the selectedPlayer will be 2. but we only have 0 and 1 as indexes since its only two players, this means that we can set the whole player selection box display as none, so we can only see the game and continue from there.