startGame() that displays icons and player names
Firstly i added the scoreBoard DOM element in the players, i did this so its easier to append the username and images later, and also in the future when im doing the scoring i can easier access each DOM element and just add a <p> for the score etc. I used this in the startGame() function that has a for loop for each player in players, doing this i can easier append text and the image on both sides instead of repeating it and having unnessecary long code. now i can take the icon from the objects which i saved before, add it to the scoreBoard on each side with appendChild, then i will create an h2 and add textContent which will be each players userName and finally append it to the scoreBoard. I will run this function directly after the choosePlayerInfoSectionDOM is none, because then the game will actually "start"