added game board, filling cols and rows
Added createGameBox which creates boxes, it treturns the box it creates as well. with this i can then use a for loop in the startGame that loops thru rows and cols, adding these to arrays and then putting in the DOM and value. With the DOM i can check which element is already used (for the each player) and if its not used, this is why i have value to store which player has taken the box. (i dont really have a solid plan for this yet, this commit is mostly cause i want to move my code to my main PC) But for now i will only create the playing main game board, and store them in someway in a variable