Hovering over gameboard now makes the first box red in the first row.
In my last commit i wrote that i made a variable that has the whole "game" in it, an array with arrays that are the rows, and then objects in them that are the different columns. I then used this to both create the boxes and also return DOM objects and values to the array. Now with this i can check multiple things, using this its easier in the code to identify which one i have hovered, clicked on etc. Here i firstly added an eventListener to my boxes, with mouseover and mouseout (basic hovering) to each box that i create. It uses the function hoverGameBox, where i send in the eventObject. With this i can first declare a constant, boxDOM, which i have the gameBoard (array) taking out the first row, (i want the hoverable effect to be on the top row always, this will only symbolize the "dropping" effect you get when you play connect 4, starting always at the top) (NOTE: this is only the hovering effect, meaning that later when i actually click it i want it to appear at the bottom). So i now have the first row in the grid, i will then use a function called findCol, which does exactly what the name is, finds the column where the DOM is that we are hovering. It takes in a boxDOM as an arg, loops thru the rows (first array) in the gameBoard, then loops thru each array in the arrays (the cols). If it finds that the DOM element in this object is the same as the element that is in the param (the one we hover on) then we can return x, which is the number of what col it is in. If we have this we can easily type out gameBoard[0] (first row) [findCol(e.target)] where findCol finds the exact column and returns its index, e.target is the current DOM we are on (the one we put the eventListener on) and then we can make a check to mark or remove the mark. Since it will only be marked once we hover on it, we can use classList.contains ("marked") first and check if it already has the class, if it does, remove it (mouseout will do this) and if it doesnt already have it, add it (mouseover will do this)

By using this array in an array with objects, i can easily modify and use the game board in a good way. the for loops in startGame() can both add the boxes on the board (using appendChild in createGameBox) and also add the DOM in the arraylist (using the return from createGameBox) . the value is something i will use later on when i click on it. I will probably put the value as player 1 or player 2