Add players CSS fixes + js code to showcase the selected icon
Added a better background with gradiant using https://cssgradient.io/ changed text color to white, added some styling to the icon images and a border. i want everything to look like an old windows computer / inspired by "frutiger areo"-style and windows 7 / windows xp. ( a little mix of everything ),

To the javascript i added a border to the images instead of the buttons, then i added some code that will create an element above the icons so you can see the one you selected but a bigger version. I will check if there is something in the innerHTML of the #player-icon-selected div, if its empty then go to the next step in the code, if its not it resets it and then goes to the next step, everytime it will update the div with a img inside of it, and reset it to the newer one you clicked on if theres already one you clicked on before (so they do not appear beside eachother)