Added Click function to the board for players to place their "box/game piece" with a background color
The first function i added was "findEmptySpotRow(col)" which finds the row of the last empty spot in the column, i will return once i have found a box that is not empty. So it goes thru the gameBoard arrays length, then it will check each rows value in the column (the param we put in) if something isn't null, that one is already taken and then the function will return the index before the one that is already occupied. In this way i can always look for the last one so i get this "drop" effect, the piece will always be at the "bottom". And ofcouse in the beginning everything will be null, therefor i can return the last index instead (if the if-statement is false).

Then i will add an eventListener to each box i create, when you click on it it will firstly find the column that it is in and i have that as a const (the same function i use on mouseover and mouseout, with e.target (the current box)) And then i will also save a const with the last rows index that is empty, with the col as an arg. Then i will use the gameBoard and put in the index row and col, and set the value in the object to the current player. Each time i click i will also change the player, i do this in a simple if statement, if the player is 0 (which i put as 0 before i run the function startGame()) then it will firsly add a classList that has a background in the css on the right box, then it will put the selectedPlayer as 1 (the next player), and then the same but the other way around.