Winner checking (Horizontal & Vertical)
added some functions to check the winner horizontally and vertically. The main function is checkWinner(col, row) this takes in the current column and row as a param, then it has some variables that checks horizontally, veritcally and diagonally. These varaibles will be arrays since each function returns an array of all boxes that made the player win. (i need to have these so i can mark them in some way to tell the player how they won). The function needs to check if any arrays are empty or if it isn't. (will continue with this later).

The checkWinnerVerti will loop thru the gameBoard with its index "row", if the gameBoards[row] and [currentCol]s value is the same as the selectedPlayer (if the objects value is the same as the selectedPlayers value), but if it isn't the same- then it should just empty the array because then the value isnt the same, (theres not 4 in a row, theres something inbetween). But if the winningArrayOfBoxes is 4, then it can break the loop and return. The same goes for the horizontal but instead we will check columns in the current row.