Button styling, function for colors on boxes and background depending on the icon
I took a break from the annoying winner check coding and did some user stuff instead;
Firstly to access the #player-icon-selected DOM i instead put it in the global scope, then i made a function called iconStylingsAndColors with a param "imageSource" which is the current clicked image's source. I use this function in the "iconSelector()", there i will put the background of the  #player-icon-selected (the main div) to the background that depends on the image you clicked. In the iconStylingsAndColors() i will add a new key and value to the object "players[selectedPlayer]" with boxColor and yellow, this will then be used to make the players have a unique color in the connect 4 game, depending on the picture you clicked. dog = yellow, bear = brown, fox = red. Then the function will also return a linear-gradient that has been modified depending on the color. (these gradients are also created using https://cssgradient.io/, like the main background). This gradient background will now be returned to the iconSelector() function and be set as the background for the whole main div.

Now the boxes also get a color, instead of putting a class on them i will directly select the current dom with the gameBoard array and its row and cols, and then just put a background color which is the value i saved in the players[selectedPlayer].boxColor

I also added some messages in the beginning, setting "welcome player 1!" directly in the code at the start, then once you have clicked the button it will say "welcome player 2!" this is just to make the user more sure whats going on.

In the css i also added some basic styling to a button making it more "old windows" which is the style im going for as i said in previous commits.