checking winner diagonally, using a for loop and if statements (see desc)
I started working on the function checkWinnerDiag() which takes in the currentCol and currentRow as args (the current column and current row from the one game tile/box that the played put out). The for loop index name is called "jumpInGrid" and then couns to three. JumpInGrid means that i will jump up and down from the current row and current col im on, this will check the diagonal boxes above the box we placed. I also need to add an if-statement to check if we are still in the grid in someway, if it counts to three but theres only 2 columns left from where you placed the box, then it cant check there because there are no boxes, it will be undefined (cant find any more objects or arrays in gameBoard). So if its not undefined, then it can continue and also check if the value of the boxes diagonally is the same as the selectedplayer, if it is, push this box into the winningArrayOfBoxes (the one array we used before which is just an array of every DOM that will later mark out how the played won). How we check the diagonal boxes is with gameBoard[currentRow-jumpInGrid][currentCol-jumpInGrid]. we are going down rows (which means we are going "up") but we are at the same time going up in columns (this means we can jump cols and not stay in the same col) this makes a diagonal check.

The problem with this is that it only checks the diagonal from the box we placed and then the diagonal "up" we also need to check the diagonal "down" from the box we placed, this function will run each time we have clicked on a box, when the played has selected where they want to place out their game tile, so from this current game tile we will with the current function check the diagonal above from the placed one.