Checking diagonally but down from the placed box instead, also checking if it works or not
I added the same for loop to check down from the box, the last commit i checked up from the one i placed, now i check down diagonally instead. I did this with just switching that im adding up the row and down the column, (the other way around)

I also added a new check to see if theres another players tile in between or if the last tile is the other players tile, because if theres only one above in the diagonal or one bellow in the diagonal down, i need to end right there, i cant continue since the other players tile is afterwards in the diagonal im checking. Thats why i need to add a check, if the upDiagonallyBox (just a quick variable so i dont need to repeat gameBoard[CurrentRow - jumpInGrid][currentCol + jumpInGrid] all the time) if that box has the same value as the current selected player, we  can push that to the winning tiles/boxes list. But we also before pushing that need to check if the previous value is NOT the same as the selectedplayer. this is because if it firstly can do the check and it goes okay it just pushes, if the next value is not the selectedplayer, it wouldnt do anything, continue, and then if the last check is the same as the selectedplayer it would add that as a winning box. This would be wrong since there would be one box that is the other players tile in between the selectedplayers tiles. making it not winable. Thats why i need to check if the previous one is not the same as the selectedplayers one. and then i could empty the list since it would not work. else it can push it into WinningArrayOfBoxes. This goes well with also checking if the last box in the diagonal is not the selectedPlayers box, because its checks the previous one. If the last box is the other players box but the previousbox is the current player, then it wont reset the list. Since its checking the previous box instead of checking like, checking the current box, it wont check the last one and empty the list if its not the selectedPlayers box. So the check goes both ways- checking if theres another players box in between and not the selectedplayer- it will empty. If the last box in the diagonal is not the selectedPlayers box, it cannot empty the array of the winning boxes, it only adds the boxes before that if they are the selectedPlayers boxes.

ofcouse before i go into the next iteration of the loop i will set the previousBoxValue to the upDiagonallyBox's value (resets it)

Then i did the same for the down diagonal, same check, just going up in row and down in columns. the PreviousBoxValue needs to be reseted as well to the selectedPlayer, and also i need to start this for loop at 1 instead (I dont want to count the starting box (the one we are checking from if we won) because i dont want it twice in my winning array of doms)