Fixed vertical not working with reseting the winningArrayOfBoxes each time
The vertical winning function would not work as i wanted because there were boxes left from the previous check. They were there because i pushed them and never removed them afterwards. If the playerSelected if-statement goes thru, they will just be pushed to the winningArrayOfBoxes, thats why i need to reset it before clicking on a new tile.

I also realized that the checkWinnerDiag needs to have two functions, one for checking the right diagonal up and down and one for checking the left diagonal up and down, thats why i added "Right" to the function