Hightlights a players turn on the board
Added a way to highlight the players turn, i do this by adding a boxshadow and also a border. using the players[(selectedplayer)]. I added the function changeTurn that takes in a param, playerToChangeTo which is the player that it will change to, selectedplayer is the current player. Then it will put none to the selectedPlayer and then add a boxShadow & border to the player that we will change to, (the next player before them clicking) ending it with setting selectedPlayer to the player we will change into. I need to put the other player in the param, this is cause i want the effect to "change" turn after you have selected your own tile. also added some other styling to the board using CSS and z-index with position relative and absolute. The border will appear behind the window, this is used to make a cool border effect. Also the game boxes and the actual game looks more like connect-4, using border-radius and background color.

++added a score h2 element that will later show the score