Optimized diagonal checker, now checks right and left in one function
I was supposed to have two different functions, one for checking the diagonal right up to left down and one for checking left up to right down. Instead of repeating the same code 4 times- i decided to test some different ways to fix it better. With using more params, sending in operators as strings to the function, i can easily check what the string is and then change the the value of how we want to jump in the grid.

In checkWinnerDiag i will use the variable "jumpRow" and "jumpCol", this uses an if-statement (i used a ternary operator instead to minimize code lines and it makes it easier to read) i will check if op1 is -, if it is, put the variable "JumpRow" as currentRow - JumpInGrid, if not put it as + instead. and op2 just does the same put with cols. I will also do this check in the second for-loop "check" of the function. But instead if it isn't -, i will put it as - and if it is - i will put it as +.

I will call the function twice, first i will put op1 as - and then op2 as +, then in the function the first if-statment will be true, putting currentRow - jumpInGrid as the "jumpRow" variable. And then the second variable, jumpCol will be false in the check, making it "currentCol + jumpInGrid" then the function will just put out the same as it was before when i wanted to make  a checkWinnerDiagRight; go down in rows but up in cols.

For the second time i call the function, the args will now be "-" and "-" making both op1 and op2 -. When checking the left instead i want to firstly go down in rows and down in columns. then i want to check when i go up in rows and up in columns. In the function, the variables "jumpRow" and "jumpCol" will now both be true in the check, returning "currentRow - jumpInGrid" aswell as "currentCol - jumpInGrid". The other check going down will be + instead, because i want to go up in cols and up in rows.

I feel like this was an okay solution, using extra params that switches the check so i can check both the right diagonal and the left diagonal.