Displays winner and stops the game!
I can now display the winner with their name at the top, show how they won on the game board and also added a reset button with css (not working yet)

Firstly i added some fixes to the border, using a box-shadow and border, also display flex with justify content flex-end to make the reset button be all the way to the right. Then i added the actual button, made the color red (i wanted it to look like a exit button :)) and then i added a height and width, (they look specific but they are random numbers) and then the text color to be white, border-radius, cursor to pointer because its a button (I did this for the game tiles as well)

going to the actual code i decided to add a new function called outPutWinner() which doesnt take in any args, it only displays a new textContent for the h2 in game-section, also it goes thru the whole list of the winningArrayOfBoxes and puts a boxshadow and a border around each box in the list (this is what i wanted to do from the beginning when i made this list!)

Then i need to be able to check each function if its a horiziontal win, vertical win, or diagonal both left and right. i do this by first declaring a variable for the function, then check if winningArrayOfBoxes length is bigger or 4 (im also putting bigger since you could get 5 and that would count as a win). Also when i check horizontal and vertical, they return winningArrayOfBoxes in their function so thats why i put a .length on the variable with the function instead. And if it is 4, run outPutWinner(), and then return. Return makes the function stop, which is nessecary because i dont want to keep on going, if it finds out that the array is 4, then just stop, dont check the others. Also something i need to do in between them all is reseting winningArrayOfBoxes, doing this i prevent it from still having box elements in it. It can find 2 boxes in one function, and then when the function has runned those 2 boxes are still in winningArrayOfBoxes, which is not good. So to prevent any mix up or "left overs" from the other functions.. i will simply just reset it just in case.

Also i dont want any more tiles after victory, game over means game over. By locking the board with using a function called "lockMode" and putting in the otherPlayer, i can firstly remove any styling that would be added because of the changeTurn function, and then also set a variable to true. I have a variable in the global scope that is always false, until i run the function "lockMode", this will run after changeTurn on each player, when winningArrayOfBoxes.length if 4.. run the function lockMode. Doing this it will also "lock" the click eventListener, because i put an if-statement directly in the click eventListener on the box element. If its false, (nobody has won and the game goes on) it will just continue doing everything like it did before. if its true, then it wont do anything when you click.