Added scoring system
Now i have a new key and value to the players array, which is wins. I changed the textContent instead to the players[i].wins, which is 0 in the beginning. then each time someone wins i will use the function outPutWinner() and in this i will increment the selectedplayer wins with 1. Then i will also display this on the textConnect of the h2 under the name.