Reset button working - added eventListener with reseting different variables and running a createTheWholeGameBoard func
I changed gameBoard to let instead of const, because i need to reset the whole thing when i click the reset button and i cant change the variable if its a constant.

then i decided to have a h3 instead of a h2 when displaying the wins for each player.

When i create the whole gameBoard i did that directly in the startGame() function, instead i wanted to make another function that creates a board with all the boxes. Doing this, i can now run the function in startGame to start it in the beginning after i choose avatar and name for the players. Then i can also run the function when i click the reset button. Moving on to when i click the reset button..

I added an eventListener to the button, when clicking it, it will firstly clear the gameBoard (so i can add all the objects again in the createTheWholeGameBoard), clear the winningArrayOfBoxes (feels odd to still have the winning boxes of the game before, it needs to be reseted), New textContent for the winnerOutputDOM (so it doesnt display who won the game before) changed lockedBecauseSomeoneWon to false, (because it changes to true when you win the game; i need to lock the whole gameboard and setting it to false again "unlocks" the board so i can click) and then i will wipe the whole game-board elements innerHTML as well, to acutally delete the whole board before. Then i will run the function createTheWholeGameBoard() and it will just add back everything because its the same function that i use when i just start the game. I will also change the turn to 0, like it is from the start.