Checking if theres a draw- the whole board is filled.
I added a function called checkDraw() that checks if the whole game board is already filled. i did this by putting the checkDraw() function after the checkWinner() function, it will check if its a draw by first looping thru the gameBoard in a nested loop, then checking if the gameBoards individual boxes has a value that isn't null, its set as null in the beginning  on each value (therefor when all the values are replaced with the different players then there wont be any null values). and then it will output a text that says "draw!". Though in the beginning there will be values that are null at the start, this means that it will always output draw when you click because the values are not filled up, but it will still find values that are null and then just output it directly cause it found a value that isn't null. But if it finds a value that is null, it will still output game on, meaning that the draw will only activate once it cant find any values with null at all. Return makes the loops end because then it sends back to the function, but i dont necessarly want to return anything, i just want a quick way to exit both loops and the whole function. Because if a value is null; it cant be full.