moved checkDraw to above checkWinner!
I needed to do this for it to be able to work. This is because it will always change back to "game off!" when using checkDraw, if winner is before, then it will still go thru checkDraw and change the textContent to "game off"! instead of it being "___ won!", thats why i needed to put it above. (also changed some variables